+++ ricaun.RevitTest.TestAdapter/Models/RunSettingsModel.cs
@@ -11,14 +11,14 @@
 
         public class NUnitModel
         {
-            [XmlElement("RevitVersion")]
-            public int RevitVersion { get; set; }
+            [XmlElement("Version")]
+            public int Version { get; set; }
 
-            [XmlElement("RevitOpen")]
-            public XmlBool RevitOpen { get; set; }
+            [XmlElement("Open")]
+            public XmlBool Open { get; set; }
 
-            [XmlElement("RevitClose")]
-            public XmlBool RevitClose { get; set; }
+            [XmlElement("Close")]
+            public XmlBool Close { get; set; }
 
             [XmlElement("Verbosity")]
             public int Verbosity { get; set; }